extract MappedFutures to module nested in FuturesUnordered; clean code
